Move CheckPointWriter to checkpoints mod
